requirements.txt
# Core dependencies
langchain==0.1.0
langchain-community==0.0.10
langchain-core

# LLM Support
openai>=1.0.0
google-generativeai>=0.3.0

# Vector store
faiss-cpu==1.7.4
# For GPU support, replace faiss-cpu with: faiss-gpu==1.7.4

# Embeddings
sentence-transformers==2.3.1
torch==2.1.2
transformers==4.36.2

# Web Interface
streamlit==1.31.0

# Utilities
numpy==1.24.3
requests>=2.28.0
